refactor(frontend): consolidate auth context files and fix favicon via Vite import
