+++ frontend/index.html
@@ -2,7 +2,6 @@
 <html lang="en">
   <head>
     <meta charset="UTF-8" />
-    <link rel="shortcut icon" href="/RA-ICON_embed.svg" />
     <meta name="viewport" content="width=device-width, initial-scale=1.0" />
     <title>Rift Analytics</title>
